Work commit.
git-svn-id: https://chrome-rest-client.googlecode.com/svn/trunk@125 4310ba71-df81-c583-98c1-67588e199800

Former-commit-id: 53de6ba975b980bc21f40249c2d0b7fdf94ae57f [formerly 415ce19d483f11c85d489049529e6fa635e5fad5]
Former-commit-id: 79b84a37bec9e657840503860366a03e9f643983

+++ RestClient/src/org/rest/client/chrome/HistoryItem.java
@@ -39,7 +39,7 @@ public class HistoryItem extends JavaScriptObject {
 	 * @optional
 	 * @return the lastVisitTime
 	 */
-	public final native long getLastVisitTime() /*-{
+	public final native double getLastVisitTime() /*-{
 		return this.lastVisitTime;
 	}-*/;
 	/**
